#156 Refactor test
